refactor(api): consolidate platform-specific routes into single api v1 structure
- Remove setEnvironment middleware and platform-specific route handlers
- Move web/auth and web/security routes directly under api/v1
- Delete empty platform route files (app, bot, desktop, wearable)
- Update main router to use unified v1 endpoint with context middlewares
- Reorganize route imports and context middleware application flow
