+++ routes/auth/index.js
