--- routes/desktop/index.js
@@ -1,14 +1,0 @@
-// =============================================================================
-// THIRD-PARTY DEPENDENCIES
-// =============================================================================
-const express = require('express');
-
-// =============================================================================
-// SET UP ROUTER
-// =============================================================================
-const router = express.Router();
-
-// =============================================================================
-// MODULE EXPORTS
-// =============================================================================
-module.exports = router;
