--- routes/web/index.js
@@ -1,33 +1,0 @@
-// =============================================================================
-// THIRD-PARTY DEPENDENCIES
-// =============================================================================
-const express = require('express');
-
-// =============================================================================
-// INTERNAL DEPENDENCIES
-// =============================================================================
-const { pageUseEndpoint } = require('../../middlewares/audit/pageEndpointLogger.middleware');
-const { setHost, setPage, setEndpoint } = require('../../middlewares/context/contextBuilder.middleware');
-
-// =============================================================================
-// SET UP ROUTER
-// =============================================================================
-const router = express.Router();
-
-router.use(setHost);
-router.use(setPage);
-router.use(setEndpoint);
-
-router.use(pageUseEndpoint);
-
-// =============================================================================
-// ROUTES
-// =============================================================================
-router.use('/auth', require('./auth'));
-
-router.use('/security', require('./security'));
-
-// =============================================================================
-// MODULE EXPORTS
-// =============================================================================
-module.exports = router;
